Authenticator well implemented with roles and logout

+++ src/main/charter/controllers.js
@@ -1,26 +1,41 @@
 import { PrismaClient } from "@prisma/client";
+import hasRoles from "../../helpers/auth.js";
 
 const prisma = new PrismaClient();
 
 export async function getOneCharter(req, res) {
-  const { id } = req.params;
+  const { roles } = req;
 
-  let cId = 0;
+  try {
+    if (!hasRoles(roles, ["ADMINTOOL", "ADMINLEAD"]))
+      return res.sendStatus(401);
+    const { id } = req.params;
 
-  if (id && !isNaN(Number(id))) {
-    cId = Number(id);
+    let cId = 0;
+
+    if (id && !isNaN(Number(id))) {
+      cId = Number(id);
+    }
+
+    const getCharter = await prisma.Charter.findUnique({
+      where: { id: cId },
+    });
+
+    res.json(getCharter);
+  } catch (err) {
+    console.log(err);
+    return res
+      .status(500)
+      .json({ error: "An error occurred while getting one Charter" });
   }
-
-  const getCharter = await prisma.Charter.findUnique({
-    where: { id: cId },
-  });
-
-  res.json(getCharter);
 }
 
 export async function createCharter(req, res) {
   const { name, projectId } = req.body;
+  const { roles } = req;
+
   try {
+    if (!hasRoles(roles, ["ADMINLEAD"])) return res.sendStatus(401);
     const newCharter = await prisma.Charter.create({
       data: {
         name,
@@ -40,7 +55,10 @@ export async function createCharter(req, res) {
 export async function updateCharter(req, res) {
   const { id } = req.params;
   const { name } = req.body;
+  const { roles } = req;
+
   try {
+    if (!hasRoles(roles, ["ADMINLEAD"])) return res.sendStatus(401);
     const newCharter = await prisma.Charter.update({
       data: {
         name,
@@ -59,7 +77,10 @@ export async function updateCharter(req, res) {
 
 export async function deleteCharter(req, res) {
   const { id } = req.params;
+  const { roles } = req;
+
   try {
+    if (!hasRoles(roles, ["ADMINLEAD"])) return res.sendStatus(401);
     const deleteCharter = await prisma.Charter.delete({
       where: { id: Number(id) },
     });
