+++ src/main/project/controllers.js
@@ -1,133 +1,180 @@
 import { PrismaClient } from "@prisma/client";
+import hasRoles from "../../helpers/auth.js";
 
 const prisma = new PrismaClient();
 
 export async function getAll(req, res) {
-  const Projects = await prisma.Project.findMany();
-  return res.json({ Projects });
+  const { roles } = req;
+
+  try {
+    if (!hasRoles(roles, ["ADMINTOOL", "ADMINLEAD"]))
+      return res.sendStatus(401);
+    const Projects = await prisma.Project.findMany();
+    return res.json({ Projects });
+  } catch (err) {
+    console.log(err);
+    return res
+      .status(500)
+      .json({ error: "An error occurred while getting the Projects" });
+  }
 }
 
 export async function getProjectTree(req, res) {
-  const ProjectTree = await prisma.Project.findMany({
-    include: {
-      Charter: true,
-      ProjectSoftwares: {
-        select: {
-          user: {
-            select: {
-              id: true,
-              name: true,
-              email: true,
+  const { roles } = req;
+
+  try {
+    if (!hasRoles(roles, ["ADMINTOOL", "ADMINLEAD"]))
+      return res.sendStatus(401);
+    const ProjectTree = await prisma.Project.findMany({
+      include: {
+        Charter: true,
+        ProjectSoftwares: {
+          select: {
+            user: {
+              select: {
+                id: true,
+                name: true,
+                email: true,
+              },
             },
-          },
-          software: {
-            include: {
-              Task: {
-                include: {
-                  Subtask: true,
+            software: {
+              include: {
+                Task: {
+                  include: {
+                    Subtask: true,
+                  },
                 },
               },
             },
           },
         },
       },
-    },
-  });
+    });
 
-  const groupedProjectSoftwares = ProjectTree.reduce((acc, project) => {
-    const softwares = project.ProjectSoftwares.reduce(
-      (acc, projectSoftware) => {
-        const key = projectSoftware.software.code;
-        const existingSoftware = acc[key];
+    const groupedProjectSoftwares = ProjectTree.reduce((acc, project) => {
+      const softwares = project.ProjectSoftwares.reduce(
+        (acc, projectSoftware) => {
+          const key = projectSoftware.software.code;
+          const existingSoftware = acc[key];
 
-        if (!projectSoftware.user) {
-          // Si no hay usuario (administrador), se crea una entrada en ProjectSoftwares sin usuarios
+          if (!projectSoftware.user) {
+            // Si no hay usuario (administrador), se crea una entrada en ProjectSoftwares sin usuarios
+            if (existingSoftware) {
+              return acc;
+            }
+
+            const newSoftware = {
+              software: projectSoftware.software,
+              users: [],
+            };
+
+            acc[key] = newSoftware;
+            return acc;
+          }
+
           if (existingSoftware) {
+            existingSoftware.users.push(projectSoftware.user);
             return acc;
           }
 
           const newSoftware = {
             software: projectSoftware.software,
-            users: [],
+            users: [projectSoftware.user],
           };
 
           acc[key] = newSoftware;
           return acc;
-        }
+        },
+        {}
+      );
 
-        if (existingSoftware) {
-          existingSoftware.users.push(projectSoftware.user);
-          return acc;
-        }
-
-        const newSoftware = {
-          software: projectSoftware.software,
-          users: [projectSoftware.user],
-        };
-
-        acc[key] = newSoftware;
-        return acc;
-      },
-      {}
-    );
-
-    const newProject = {
-      ...project,
-      ProjectSoftwares: Object.values(softwares),
-    };
-    acc.push(newProject);
-    return acc;
-  }, []);
-
-  res.json(groupedProjectSoftwares);
+      const newProject = {
+        ...project,
+        ProjectSoftwares: Object.values(softwares),
+      };
+      acc.push(newProject);
+      return acc;
+    }, []);
+    res.json(groupedProjectSoftwares);
+  } catch (err) {
+    console.log(err);
+    return res
+      .status(500)
+      .json({ error: "An error occurred while getting the Projects Tree" });
+  }
 }
 
 export async function getOneProject(req, res) {
   const { id } = req.params;
+  const { roles } = req;
 
-  let pId = 0;
+  try {
+    if (!hasRoles(roles, ["ADMINTOOL", "ADMINLEAD"]))
+      return res.sendStatus(401);
+    let pId = 0;
 
-  if (id && !isNaN(Number(id))) {
-    pId = Number(id);
+    if (id && !isNaN(Number(id))) {
+      pId = Number(id);
+    }
+
+    const getProject = await prisma.Project.findUnique({
+      where: { id: pId },
+    });
+
+    res.json(getProject);
+  } catch (err) {
+    console.log(err);
+    return res
+      .status(500)
+      .json({ error: "An error occurred while getting one Project" });
   }
-
-  const getProject = await prisma.Project.findUnique({
-    where: { id: pId },
-  });
-
-  res.json(getProject);
 }
 
 export async function getAdminAndSoftwareFromProject(req, res) {
-  const { adminId, softwareId, projectId } = req.params;
+  const { roles } = req;
 
-  let pId = 0;
-  let sId = 0;
-  let aId = 0;
+  try {
+    if (!hasRoles(roles, ["ADMINTOOL", "ADMINLEAD"]))
+      return res.sendStatus(401);
+    const { adminId, softwareId, projectId } = req.params;
 
-  if (projectId && !isNaN(parseInt(projectId))) {
-    pId = parseInt(projectId);
+    let pId = 0;
+    let sId = 0;
+    let aId = 0;
+
+    if (projectId && !isNaN(parseInt(projectId))) {
+      pId = parseInt(projectId);
+    }
+    if (softwareId && !isNaN(parseInt(softwareId))) {
+      sId = parseInt(softwareId);
+    }
+    if (adminId && !isNaN(parseInt(adminId))) {
+      aId = parseInt(adminId);
+    }
+    const getAdminSoftwares = await prisma.ProjectSoftwares.findMany({
+      where: {
+        adminId: aId,
+        softwareId: sId,
+        projectId: pId,
+      },
+      select: { id: true },
+    });
+    res.json(getAdminSoftwares);
+  } catch (err) {
+    console.log(err);
+    return res.status(500).json({
+      error:
+        "An error occurred while getting an Admin and Software from Project",
+    });
   }
-  if (softwareId && !isNaN(parseInt(softwareId))) {
-    sId = parseInt(softwareId);
-  }
-  if (adminId && !isNaN(parseInt(adminId))) {
-    aId = parseInt(adminId);
-  }
-  const getAdminSoftwares = await prisma.ProjectSoftwares.findMany({
-    where: {
-      adminId: aId,
-      softwareId: sId,
-      projectId: pId,
-    },
-    select: { id: true },
-  });
-  res.json(getAdminSoftwares);
 }
 
 export async function createProject(req, res) {
   const { name, code, estimatedHours, userProjectId } = req.body;
+  const { roles } = req;
+
   try {
+    if (!hasRoles(roles, ["ADMINLEAD"])) return res.sendStatus(401);
     const newProject = await prisma.Project.create({
       data: {
         name,
@@ -157,7 +204,10 @@ export async function createProject(req, res) {
 
 export async function addSoftwareAndAdminToProject(req, res) {
   const { projectId, adminId, softwareId } = req.body;
+  const { roles } = req;
+
   try {
+    if (!hasRoles(roles, ["ADMINLEAD"])) return res.sendStatus(401);
     const newSoftwareProject = await prisma.ProjectSoftwares.create({
       data: {
         projectId: Number(projectId),
@@ -197,7 +247,10 @@ export async function addSoftwareAndAdminToProject(req, res) {
 export async function updateProject(req, res) {
   const { id } = req.params;
   const { name, code, estimatedHours } = req.body;
+  const { roles } = req;
+
   try {
+    if (!hasRoles(roles, ["ADMINLEAD"])) return res.sendStatus(401);
     const newProject = await prisma.Project.update({
       data: {
         name,
@@ -218,7 +271,10 @@ export async function updateProject(req, res) {
 
 export async function deleteProject(req, res) {
   const { id } = req.params;
+  const { roles } = req;
+
   try {
+    if (!hasRoles(roles, ["ADMINLEAD"])) return res.sendStatus(401);
     const deleteProject = await prisma.Project.delete({
       where: { id: Number(id) },
     });
@@ -234,7 +290,10 @@ export async function deleteProject(req, res) {
 
 export async function removeAdminSoftware(req, res) {
   const { id } = req.params;
+  const { roles } = req;
+
   try {
+    if (!hasRoles(roles, ["ADMINLEAD"])) return res.sendStatus(401);
     const removeAdminSoftware = await prisma.ProjectSoftwares.update({
       where: { id: Number(id) },
       data: {
